Move some items from server/kv.rs to storage
