+++ components/test_coprocessor/src/store.rs
@@ -9,7 +9,11 @@ use kvproto::kvrpcpb::{Context, IsolationLevel};
 use test_storage::{SyncTestStorage, SyncTestStorageBuilder};
 use tidb_query::codec::{datum, table, Datum};
 use tidb_query::expr::EvalContext;
-use tikv::storage::{txn::FixtureStore, Engine, RocksEngine, SnapshotStore, TestEngineBuilder};
+use tikv::storage::{
+    kv::{Engine, RocksEngine, TestEngineBuilder},
+    txn::FixtureStore,
+    SnapshotStore,
+};
 use tikv_util::collections::HashMap;
 use txn_types::{Key, Mutation, TimeStamp};
