+++ src/storage/mod.rs
@@ -10,12 +10,12 @@
 //! [`RocksEngine`](storage::RocksEngine) are used for testing only.
 
 pub mod config;
+pub mod errors;
 pub mod kv;
 pub mod lock_manager;
 pub mod mvcc;
 pub mod txn;
 
-mod errors;
 mod metrics;
 mod readpool_impl;
 mod types;
@@ -28,7 +28,7 @@ pub use self::{
     },
     readpool_impl::{build_read_pool, build_read_pool_for_test},
     txn::{Options, ProcessResult, Scanner, SnapshotStore, Store},
-    types::{MvccInfo, StorageCallback, TxnStatus},
+    types::{StorageCallback, TxnStatus},
 };
 
 use crate::storage::{
@@ -41,6 +41,7 @@ use crate::storage::{
         scheduler::Scheduler as TxnScheduler,
         PointGetCommand,
     },
+    types::MvccInfo,
 };
 use engine::{CfName, IterOption, ALL_CFS, CF_DEFAULT, DATA_CFS, DATA_KEY_PREFIX_LEN};
 use futures::{future, Future};
