+++ tests/integrations/storage/test_raft_storage.rs
@@ -9,10 +9,10 @@ use std::sync::Arc;
 use test_raftstore::*;
 use test_storage::*;
 use tikv::server::gc_worker::{AutoGcConfig, GcConfig};
-use tikv::storage::kv::{Error as KvError, ErrorInner as KvErrorInner};
+use tikv::storage::kv::{Engine, Error as KvError, ErrorInner as KvErrorInner};
 use tikv::storage::mvcc::{Error as MvccError, ErrorInner as MvccErrorInner};
 use tikv::storage::txn::{Error as TxnError, ErrorInner as TxnErrorInner};
-use tikv::storage::{Engine, Error as StorageError, ErrorInner as StorageErrorInner};
+use tikv::storage::{Error as StorageError, ErrorInner as StorageErrorInner};
 use tikv_util::collections::HashMap;
 use tikv_util::HandyRwLock;
 use txn_types::{Key, Mutation, TimeStamp};
